Fix tests for numpy 2

+++ tests/test_ensembles.py
@@ -288,8 +288,14 @@ class TestEnsembleStats:
         )
         out2 = ensembles.ensemble_mean_std_max_min(ens, weights=weights)
         values = ens["tg_mean"][:, 0, 5, 5]
+        # Explicit float64 so numpy does the expected datatype promotion (change in numpy 2)
         np.testing.assert_array_equal(
-            (values[0] * 1 + values[1] * 0.1 + values[2] * 3.5 + values[3] * 5)
+            (
+                values[0] * np.float64(1)
+                + values[1] * np.float64(0.1)
+                + values[2] * np.float64(3.5)
+                + values[3] * np.float64(5)
+            )
             / np.sum(weights),
             out2.tg_mean_mean[0, 5, 5],
         )
